[#208] SpectralLines.plot() operational and in unit tests + generic plot in DataCollection
